Rename ProcessCommandExecutor with SimpleCommandExecutor
