+++ digdag-standards/src/main/java/io/digdag/standards/operator/RbOperatorFactory.java
@@ -19,12 +19,11 @@ import com.google.common.io.CharStreams;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.inject.Inject;
 import io.digdag.spi.OperatorContext;
-import io.digdag.standards.command.ProcessCommandExecutor;
+import io.digdag.standards.command.SimpleCommandExecutor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import io.digdag.spi.CommandExecutor;
 import io.digdag.spi.CommandLogger;
-import io.digdag.spi.TaskRequest;
 import io.digdag.spi.TaskResult;
 import io.digdag.spi.Operator;
 import io.digdag.spi.OperatorFactory;
@@ -149,7 +148,7 @@ public class RbOperatorFactory
 
             // Set up process environment according to env config. This can also refer to secrets.
             Map<String, String> env = pb.environment();
-            ProcessCommandExecutor.collectEnvironmentVariables(env, context.getPrivilegedVariables());
+            SimpleCommandExecutor.collectEnvironmentVariables(env, context.getPrivilegedVariables());
 
             Process p = exec.start(workspace.getPath(), request, pb);
